small refactor, start on allows
